refactor: move validation schemas to dedicated file
