+++ utils/api/validationSchema.js
@@ -1,0 +1,25 @@
+import { truthValue } from './share'
+import { object, string, number, date, boolean, mixed } from 'yup'
+
+export const feedbackFormSchema = object({
+  name: string().required(),
+  form: string().required(),
+  ip: string().required(),
+  responseTime: date().required(),
+  field: number().positive().integer().required(),
+  userFeedback: mixed().required(),
+})
+
+export const likeFormSchema = object({
+  name: string().required(),
+  form: string().required(),
+  ip: string().required(),
+  responseTime: date().required(),
+  field: number().positive().integer().required(),
+  userFeedback: boolean()
+    .transform((value) => {
+      if (truthValue.includes(value)) return true
+      else return false
+    })
+    .required(),
+})
